implementation of fincremental and bulk deployments
